cleaning 3 (wip)

+++ src/libslic3r/Geometry/Circle.cpp
@@ -207,7 +207,7 @@ bool ArcCircle::try_create_circle(const Points& points, const double max_radius,
     double least_deviation;
     bool found_circle = false;
     double current_deviation;
-    for (int index = 1; index < count - 1; index++)
+    for (size_t index = 1; index < count - 1; index++)
     {
         if (index == middle_index)
             // BBS: We already checked this one, and it failed. don't need to do again
@@ -290,7 +290,7 @@ bool ArcCircle::get_deviation_sum_squared(const Points& points, const double tol
     Point temp;
     double distance_from_center, deviation;
     // BBS: skip the first and last points since they are on the circle
-    for (int index = 1; index < points.size() - 1; index++)
+    for (size_t index = 1; index < points.size() - 1; index++)
     {
         //BBS: make sure the length from the center of our circle to the test point is 
         // at or below our max distance.
@@ -304,7 +304,7 @@ bool ArcCircle::get_deviation_sum_squared(const Points& points, const double tol
     }
     Point closest_point;
     //BBS: check the point perpendicular from the segment to the circle's center
-    for (int index = 0; index < points.size() - 1; index++)
+    for (size_t index = 0; index < points.size() - 1; index++)
     {
         if (get_closest_perpendicular_point(points[index], points[(size_t)index + 1], center, closest_point)) {
             temp = closest_point - center;
